+++ src/docs/changelog.md
@@ -10,6 +10,7 @@ Changelog                           {#changelog}
 - removed widget example since Mica covers all you need to know about widgets in Igor
 - #374 using float32 for mouse position to reduce conversions
 - #408 cleaned up ompf from legacy data and broke compatibility with older ompf (set ompf to 2.0.0)
+- #404 made project a singleton and put project and user search paths on top in iResourceManager
 
 0.44.0
 ------